+++ CHANGES.rst
@@ -53,6 +53,8 @@ Release History
 - Nengo tests now use the ``allclose`` fixture from ``pytest-allclose``,
   which makes it possible for backends to change test tolerances.
   (`#1563 <https://github.com/nengo/nengo/pull/1563>`__)
+- Nengo tests now use the ``rng`` and ``seed`` fixtures from ``pytest-rng``.
+  (`#1566 <https://github.com/nengo/nengo/pull/1566>`__)
 
 **Changed**
 
@@ -181,6 +183,9 @@ Release History
   (`#1187 <https://github.com/nengo/nengo/pull/1187>`_)
 - ``PES.correction`` is no longer probeable.
   (`#1398 <https://github.com/nengo/nengo/pull/1398>`_)
+- The internal ``rng`` and ``seed`` fixtures have been removed. Use the
+  external `pytest-rng <https://www.nengo.ai/pytest-rng/>`__ package instead.
+  (`#1566 <https://github.com/nengo/nengo/pull/1566>`__)
 
 **Fixed**